Change every function pointer type to an Option
Rust's function pointer types do not permit null values, so constructing
one from null (with transmute, or by being passed a null pointer over
FFI) results in undefined behavior. Nullable function pointer types
should be represented as an `Option<fn(...) -> ...>`, which has the same
ABI as `fn(...) -> ...` but allows null values.

For well-behaved CLAP hosts and plugins, this should not be an issue,
since none of the function pointers found in the API are ever permitted
to be null. However, hosts and plugins are not always well behaved, and
it's desirable to be able to explicitly handle null function pointers
without invoking UB, at least in the setting of a plugin validator or
similar. So, this change wraps every function pointer type in an
`Option`.